# Portabrasil-server/docs/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PortaBrasil Flask 后端接口",
    "version": "0.1.0",
    "description": "PortaBrasil PDF 单据上传、解析入库、业务查询和费用明细接口。可直接导入 Apifox。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5001",
      "description": "本地开发服务"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "账号登录",
        "description": "支持 username 或 email 登录，校验密码（SHA-256）后返回 JWT。",
        "operationId": "authLogin",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "examples": {
                "byUsername": {
                  "summary": "用户名登录",
                  "value": {
                    "username": "admin",
                    "password": "admin123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "公开注册",
        "description": "注册新用户并自动分配默认角色（由 DEFAULT_REGISTER_ROLE 控制，默认 FORWARDER），成功后直接返回 JWT。",
        "operationId": "authRegister",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "examples": {
                "sample": {
                  "summary": "注册示例",
                  "value": {
                    "username": "newuser",
                    "password": "123456",
                    "email": "newuser@example.com",
                    "real_name": "新用户"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "忘记密码重置",
        "description": "通过用户名+邮箱重置密码（内网简化版本）。",
        "operationId": "authForgotPassword",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              },
              "examples": {
                "sample": {
                  "summary": "重置示例",
                  "value": {
                    "username": "admin",
                    "email": "admin@portabrasil.local",
                    "new_password": "newpass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重置成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "密码重置成功，请使用新密码登录"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "获取当前登录用户",
        "description": "根据 Bearer Token 返回当前用户信息和角色。",
        "operationId": "authMe",
        "responses": {
          "200": {
            "description": "当前用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/auth/users": {
      "post": {
        "tags": ["Auth"],
        "summary": "管理员创建用户",
        "description": "仅 SUPER_ADMIN/ADMIN 可调用。创建用户并分配角色。",
        "operationId": "createUserByAdmin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              },
              "examples": {
                "sample": {
                  "summary": "创建用户示例",
                  "value": {
                    "username": "zhangsan",
                    "password": "123456",
                    "real_name": "张三",
                    "email": "zhangsan@example.com",
                    "role_codes": ["CUSTOMS"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
    "/api/documents/from-text": {
      "post": {
        "tags": ["Documents"],
        "summary": "解析文本直接入库",
        "description": "把智谱返回的 content 或已有 OCR 文本直接解析并写入数据库。适合先绕过 PDF 上传流程测试规则解析。",
        "operationId": "createDocumentFromText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentFromTextRequest"
              },
              "examples": {
                "sample": {
                  "summary": "最小示例",
                  "value": {
                    "raw_text": "LOGIMEX COMERCIO EXTERIOR LTDA ... S/Ref.: 20250528000158 ... TOTAL => 228.389,13 200.897,37 ...",
                    "file_id": null,
                    "source_page_no": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "业务记录写入成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessDetailResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/business": {
      "get": {
        "tags": ["Business"],
        "summary": "查询业务列表",
        "description": "分页查询 customs_business。q 会匹配 S/Ref、Invoice、客户名称、MAWB/MBL。",
        "operationId": "listBusiness",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "搜索关键字",
            "schema": {
              "type": "string",
              "example": "20250528000158"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "每页数量，最大 100",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "偏移量",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "业务列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/business/{business_id}": {
      "get": {
        "tags": ["Business"],
        "summary": "查询业务详情",
        "description": "查询单条 customs_business，并带出 fee_items。",
        "operationId": "getBusinessDetail",
        "parameters": [
          {
            "$ref": "#/components/parameters/BusinessId"
          }
        ],
        "responses": {
          "200": {
            "description": "业务详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessDetailResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/business/{business_id}/fees": {
      "get": {
        "tags": ["Business"],
        "summary": "查询业务费用明细",
        "description": "查询指定业务下的 customs_business_fee_item 列表。",
        "operationId": "listBusinessFees",
        "parameters": [
          {
            "$ref": "#/components/parameters/BusinessId"
          }
        ],
        "responses": {
          "200": {
            "description": "费用明细列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeListResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在请求头中添加：Authorization: Bearer <access_token>"
      }
    },
    "parameters": {
...
      "BusinessId": {
        "name": "business_id",
        "in": "path",
        "required": true,
        "description": "customs_business.id",
        "schema": {
          "type": "integer",
          "example": 1
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "badRequest": {
                "value": {
                  "error": "缺少 multipart 字段 file"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未登录或 Token 无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "error": "缺少或无效 Authorization Bearer Token"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "权限不足",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "forbidden": {
                "value": {
                  "error": "权限不足"
                }
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "资源冲突（如用户名已存在）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "conflict": {
                "value": {
                  "error": "用户名已存在"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "notFound": {
                "value": {
                  "error": "业务记录不存在"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "服务端错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missingKey": {
                "value": {
                  "error": "ZHIPU_API_KEY 未配置，无法调用智谱 PDF 解析接口"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "real_name": {
            "type": "string",
            "nullable": true,
            "example": "系统管理员"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "13800000000"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "admin@portabrasil.local"
          },
          "status": {
            "type": "integer",
            "enum": [0, 1],
            "example": 1
          },
          "created_at": {
            "type": "string",
            "example": "2026-04-16 12:30:00"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-04-16 12:30:00"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "SUPER_ADMIN"
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名，username/email/account 三选一",
            "example": "admin"
          },
          "email": {
            "type": "string",
            "description": "邮箱，username/email/account 三选一",
            "example": "admin@portabrasil.local"
          },
          "account": {
            "type": "string",
            "description": "账号别名字段，等同 username/email",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "admin123456"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "newuser"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "real_name": {
            "type": "string",
            "nullable": true,
            "example": "新用户"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "13800000000"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "newuser@example.com"
          }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": ["username", "email", "new_password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "email": {
            "type": "string",
            "example": "admin@portabrasil.local"
          },
          "new_password": {
            "type": "string",
            "example": "newpass123"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": ["username", "password", "role_codes"],
        "properties": {
          "username": {
            "type": "string",
            "example": "zhangsan"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "real_name": {
            "type": "string",
            "nullable": true,
            "example": "张三"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "13800000000"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "zhangsan@example.com"
          },
          "status": {
            "type": "integer",
            "enum": [0, 1],
            "default": 1,
            "example": 1
          },
          "role_codes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["SUPER_ADMIN", "ADMIN", "FORWARDER", "CUSTOMS", "FINANCE"]
            },
            "example": ["CUSTOMS"]
          }
        }
      },
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "example": 7200
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "required": ["access_token", "token_type", "expires_in", "user"]
      },
      "CurrentUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserProfile"
          }
        },
        "required": ["user"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "错误信息"
          }
        },
        "required": ["error"]
      },
...
      "FeeItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "business_id": {
            "type": "integer",
            "example": 1
          },
          "fee_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-06-26"
          },
          "fee_code": {
            "type": "string",
            "nullable": true,
            "example": "112"
          },
          "fee_name": {
            "type": "string",
            "example": "CAPATAZIA"
          },
          "credit_amount": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "debit_amount": {
            "type": "string",
            "nullable": true,
            "example": "600.00"
          },
          "line_no": {
            "type": "integer",
            "nullable": true,
            "example": 1
          },
          "raw_text": {
            "type": "string",
            "nullable": true,
            "example": "26.06.2025 112 CAPATAZIA 600,00"
          },
          "created_time": {
            "type": "string",
            "example": "2026-04-15 22:02:00"
          }
        }
      },
      "CustomsBusinessWithFees": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CustomsBusiness"
          },
          {
            "type": "object",
            "properties": {
              "fee_items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FeeItem"
                }
              }
            }
          }
        ]
      },
      "CreateDocumentFromTextRequest": {
        "type": "object",
        "required": ["raw_text"],
        "properties": {
          "raw_text": {
            "type": "string",
            "description": "PDF 解析后的完整文本。也兼容 content 或 text 字段。",
            "example": "LOGIMEX COMERCIO EXTERIOR LTDA ... S/Ref.: 20250528000158 ..."
          },
          "file_id": {
            "type": "integer",
            "nullable": true,
            "description": "可选，关联 pdf_file.id",
            "example": 1
          },
          "source_page_no": {
            "type": "integer",
            "nullable": true,
            "description": "可选，来源页码",
            "example": 1
          }
        }
      },
...
      "BusinessListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomsBusiness"
            }
          },
          "total": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "offset": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "BusinessDetailResponse": {
        "type": "object",
        "properties": {
          "business": {
            "$ref": "#/components/schemas/CustomsBusinessWithFees"
          }
        }
      },
      "FeeListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeeItem"
            }
          },
          "total": {
            "type": "integer",
            "example": 21
          }
        }
      },
...
    }
  }
}